Add blastp module
